+++ src/argilla/client/feedback/dataset/remote/mixins.py
@@ -12,7 +12,6 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-import warnings
 from typing import TYPE_CHECKING, Iterator, List, Union
 
 from argilla.client.feedback.constants import FETCHING_BATCH_SIZE
@@ -21,13 +20,13 @@ from argilla.client.sdk.users.models import UserRole
 from argilla.client.utils import allowed_for_roles
 
 if TYPE_CHECKING:
-    from argilla.client.feedback.dataset.remote.base import RemoteFeedbackRecordsBase
+    from argilla.client.feedback.dataset.remote.dataset import RemoteFeedbackRecords
 
 
 class ArgillaRecordsMixin:
     @allowed_for_roles(roles=[UserRole.owner, UserRole.admin])
     def __getitem__(
-        self: "RemoteFeedbackRecordsBase", key: Union[slice, int]
+        self: "RemoteFeedbackRecords", key: Union[slice, int]
     ) -> Union["RemoteFeedbackRecord", List["RemoteFeedbackRecord"]]:
         """Returns the record(s) at the given index(es) from Argilla.
 
@@ -82,6 +81,8 @@ class ArgillaRecordsMixin:
         else:
             raise TypeError("Only `int` and `slice` are supported as index.")
 
+        question_id_to_name = {question.id: question.name for question in self.dataset.questions}
+
         records = []
         for offset, limit in zip(offsets, limits):
             fetched_records = self._fetch_records(offset=offset, limit=limit)
@@ -89,9 +90,7 @@ class ArgillaRecordsMixin:
                 break
             records.extend(
                 [
-                    RemoteFeedbackRecord.from_api(
-                        record, question_id_to_name=self._question_id_to_name, client=self._client
-                    )
+                    RemoteFeedbackRecord.from_api(record, question_id_to_name=question_id_to_name, client=self._client)
                     for record in fetched_records.items
                 ]
             )
@@ -102,10 +101,10 @@ class ArgillaRecordsMixin:
         return records[0] if isinstance(key, int) else records
 
     @allowed_for_roles(roles=[UserRole.owner, UserRole.admin])
-    def __iter__(
-        self: "RemoteFeedbackRecordsBase",
-    ) -> Iterator["RemoteFeedbackRecord"]:
+    def __iter__(self: "RemoteFeedbackRecords") -> Iterator["RemoteFeedbackRecord"]:
         """Iterates over the `FeedbackRecord`s of the current `FeedbackDataset` in Argilla."""
+        question_id_to_name = {question.id: question.name for question in self.dataset.questions}
+
         current_batch = 0
         while True:
             batch = self._fetch_records(offset=FETCHING_BATCH_SIZE * current_batch, limit=FETCHING_BATCH_SIZE)
@@ -113,7 +112,7 @@ class ArgillaRecordsMixin:
                 break
             for record in batch.items:
                 yield RemoteFeedbackRecord.from_api(
-                    record, question_id_to_name=self._question_id_to_name, client=self._client
+                    record, question_id_to_name=question_id_to_name, client=self._client
                 )
             current_batch += 1